ml_matcher/requirements.txt
flask==2.0.1
werkzeug==2.0.1
PyPDF2==2.10.5
python-docx==0.8.11
nltk==3.7
scikit-learn==1.0.2
numpy==1.22.3
pymongo==4.5.0
beautifulsoup4==4.11.1
requests==2.28.1
schedule==1.1.0
gunicorn==20.1.0
pytest==7.0.0
pytest-cov==2.12.1